pyproject.toml
[project]
name = "munich-sushi-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastmcp>=2.8.1",
    "flask>=3.1.1",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0",
    "google-maps-reviews>=0.0.3",
    "googlemaps>=4.10.0",
    "jq>=1.7",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "langmem>=0.0.27",
    "ollama>=0.5.1",
    "streamlit>=1.46.0",
    "tavily-python>=0.7.7",
]
